+++ src/main/java/com/klikli_dev/occultism/common/blockentity/DimensionalBattlefieldBlockEntity.java
@@ -86,7 +86,7 @@ import java.util.Collection;
 import java.util.Objects;
 import java.util.function.Consumer;
 
-public class DimensionalBattlefieldBlockEntity extends NetworkedBlockEntity implements MenuProvider {
+public class DimensionalBattlefieldBlockEntity extends NetworkedBlockEntity implements MenuProvider, Clearable {
 
     private final float BUTCHER_HURT_CHANCE = (float) Occultism.SERVER_CONFIG.itemSettings.butcherHurtChance.getAsDouble();
     private static final int DEFAULT_MAX_TIME = 20 * 20 * 20;
@@ -559,6 +559,13 @@ public class DimensionalBattlefieldBlockEntity extends NetworkedBlockEntity impl
         }
 
         return this.outputHandler;
+    }
+
+    @Override
+    public void clearContent() {
+        for (int i = 0; i < combinedHandler.getSlots(); i++) {
+            combinedHandler.setStackInSlot(i, ItemStack.EMPTY);
+        }
     }
 
     // region Inner Classes